requirements.txt
tensorboard
pytorch
tqdm
einops
scipy
torchsummary
numpy